requirements/base.txt
# Common
pyyaml

# Audio capture
numpy
scipy
aubio

# Light server
DmxPy

# Web GUI
flask
flask-bootstrap
flask-sockets